+++ packages/types-database/src/types/provider/userAccess/ip/userIpV4Mask.ts
@@ -1,13 +1,12 @@
 import { Schema } from "mongoose";
-import { Int32 } from "mongodb";
 
 // ipV4 mask is usually present like this: '198.51.100.14/24'
 // but since IP has its numeric presentation, and the mask is a rigid range,
 // we also store mask as rangeMin and rangeMax, to allow efficient greater/lesser query comparisons.
 
 interface UserIpV4Mask {
-	rangeMinAsNumeric: Int32;
-	rangeMaxAsNumeric: Int32;
+	rangeMinAsNumeric: bigint;
+	rangeMaxAsNumeric: bigint;
 	// CIDR prefix https://en.wikipedia.org/wiki/Classless_Inter-Domain_Routing - 198.51.100.14/{24}
 	// for presentation only purposes
 	asNumeric: number;
@@ -15,8 +14,10 @@ interface UserIpV4Mask {
 
 const userIpV4MaskRecordSchema = new Schema<UserIpV4Mask>(
 	{
-		rangeMinAsNumeric: { type: Int32, required: true },
-		rangeMaxAsNumeric: { type: Int32, required: true },
+		// Type choice note: Int32 can't store 10 digits of the numeric presentation of ipV4,
+		// so we use BigInt, which is supported by Mongoose and turned into Mongo's Long (Int64)
+		rangeMinAsNumeric: { type: BigInt, required: true },
+		rangeMaxAsNumeric: { type: BigInt, required: true },
 		asNumeric: { type: Number, required: true },
 	},
 	{ _id: false },
